worked on games page and extended the seed file
