+++ App.js
@@ -1,4 +1,3 @@
-
 import React from 'react';
 import { Image, StyleSheet, Text, View } from 'react-native';
 import logo from './assets/logo.png';
@@ -16,7 +15,6 @@ export default function App() {
   return (
     <React.Fragment>
       <Provider store={store}>
-        <HomePage />
 	<Navigator />
       </Provider>
     </React.Fragment>
