+++ server/db/models/index.js
@@ -4,13 +4,14 @@ const Picture = require('./picture');
 // const CluePicture = require('./cluePicture')
 const Clue = require('./clue');
 const Location = require('./location');
+const Game = require('./game');
 
-/**
- * If we had any associations to make, this would be a great place to put them!
- * ex. if we had another model called BlogPost, we might say:
- *
- *    BlogPost.belongsTo(User)
- */
+Game.belongsToMany(Clue, {
+    through: 'GameClue'
+})
+Clue.belongsToMany(Game, {
+    through: 'GameClue'
+})
 
 // User.belongsToMany(Team, {
 //   through: 'UserTeam'
@@ -19,21 +20,17 @@ const Location = require('./location');
 // through: 'UserTeam'
 // });
 
-// Picture.belongsTo(Clue);
-// Clue.hasMany(Picture);
+//Picture.belongsTo(Clue);
+//Clue.hasMany(Picture);
+
 // Picture.belongsTo(User);
 // User.hasMany(Picture);
 
-/**
- * We'll export all of our models here, so that any time a module needs a model,
- * we can just require it from 'db/models'
- * for example, we can say: const {User} = require('../db/models')
- * instead of: const User = require('../db/models/user')
- */
 module.exports = {
-	Location, 
-	Picture,
+  Location, 
+  Picture,
   User,
-//  Team,
-  Clue
+    //  Team,
+    Game,
+  Clue,
 }
